backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
sqlalchemy
psycopg[binary]==3.2.3
boto3
httpx==0.27.2
pydantic[email]
PyJWT==2.9.0
